+++ crates/notebook/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "jupyter"
+name = "notebook"
 version = "0.1.0"
 edition = "2021"
 publish = false
@@ -8,7 +8,7 @@ publish = false
 workspace = true
 
 [lib]
-path = "src/jupyter.rs"
+path = "src/notebook.rs"
 doctest = false
 
 [dependencies]
