+++ crates/notebook/Pipfile
