+++ crates/notebook/src/common.rs
@@ -1,0 +1,7 @@
+use serde::de::{self, DeserializeOwned, DeserializeSeed, Error, Visitor};
+
+pub(crate) fn parse_value<T: DeserializeOwned, E: serde::de::Error>(
+    val: serde_json::Value,
+) -> Result<T, E> {
+    serde_json::from_value::<T>(val).map_err(|err| E::custom(err.to_string()))
+}
