+++ crates/zed/Cargo.toml
@@ -55,7 +55,7 @@ inline_completion_button.workspace = true
 install_cli.workspace = true
 isahc.workspace = true
 journal.workspace = true
-jupyter.workspace = true
+notebook.workspace = true
 language.workspace = true
 language_selector.workspace = true
 language_tools.workspace = true
